Rename home page to HomePage
